requirements.txt
asgiref==3.7.2
attrs==23.2.0
backports.zoneinfo==0.2.1
certifi==2024.2.2
cffi==1.16.0
Django==4.2.8
django-autoslug==1.9.9
et-xmlfile==1.1.0
exceptiongroup==1.2.0
h11==0.14.0
idna==3.6
openpyxl==3.1.2
outcome==1.3.0.post0
Pillow==10.1.0
pycparser==2.21
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
PySocks==1.7.1
selenium==4.18.1
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.2.1
wsproto==1.2.0
